HV-1543 Split ConstraintTree into different kind of nodes
- split the logic in constraint tree into composing and non composing trees
so that in case of non composing constraint additional checks can be skipped
